+++ prayerapp/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class CardboxConfig(AppConfig):
+class PrayerappConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "cardbox"
+    name = "prayerapp"
