+++ prayerapp/templates/cardbox/index.html
@@ -3,7 +3,7 @@
 {% if usercard_list %}
     <ul>
     {% for usercard in usercard_list %}
-        <li><a href="{% url 'cardbox:usercard_detail' usercard.id %}">{{ usercard.card.title }}</a></li>
+        <li><a href="{% url 'prayerapp:usercard_detail' usercard.id %}">{{ usercard.card.title }}</a></li>
     {% endfor %}
     </ul>
 {% else %}
